test(contract): factor out shared gen_authenticated_participants helper
